Fixed auth handlers and grouped them in a seperate file
