+++ index.js
@@ -5,9 +5,12 @@ import {
   adjustForMissingHash,
   setActiveLink,
   renderTemplate,
-  fetchClient,
-  ensureProtectedRoute,
 } from "./utils.js";
+
+import {
+  ensureNotLoggedIn,
+  ensureProtectedRoute,
+} from "./auth/auth_handlers.js";
 
 import Searcher from "./models/Searcher.js";
 import "./setup/nav_setup.js";
@@ -17,7 +20,7 @@ import { initBookLists } from "./pages/booklists/index.js";
 import { initBooks } from "./pages/books/index.js";
 import { initLogin } from "./pages/login/index.js";
 import { initSignup } from "./pages/signup/index.js";
-import { loginEvent, logoutEvent } from "./setup/nav_setup.js";
+import { initNavLoginButtons } from "./setup/nav_setup.js";
 
 window.addEventListener("load", async () => {
   const bookTemplate = await loadTemplate("./pages/book/index.html");
@@ -37,11 +40,7 @@ window.addEventListener("load", async () => {
   );
 
   // Handle user login visuals
-  if (!localStorage.getItem("token")) {
-    window.dispatchEvent(logoutEvent);
-  } else {
-    window.dispatchEvent(loginEvent);
-  }
+  initNavLoginButtons();
 
   adjustForMissingHash();
   router
@@ -66,13 +65,25 @@ window.addEventListener("load", async () => {
           before: async (done) => ensureProtectedRoute(done),
         },
       },
-      "/login": () => {
-        renderTemplate(loginTemplate, "content");
-        initLogin();
+      "/login": {
+        as: "login",
+        uses: () => {
+          renderTemplate(loginTemplate, "content");
+          initLogin();
+        },
+        hooks: {
+          before: async (done) => ensureNotLoggedIn(done),
+        },
       },
-      "/signup": () => {
-        renderTemplate(signupTemplate, "content");
-        initSignup();
+      "/signup": {
+        as: "signup",
+        uses: () => {
+          renderTemplate(signupTemplate, "content");
+          initSignup();
+        },
+        hooks: {
+          before: async (done) => ensureNotLoggedIn(done),
+        },
       },
       "/book/:bookId": (param) => {
         renderTemplate(bookTemplate, "content");
